+++ firebase-dataconnect/src/main/kotlin/com/google/firebase/dataconnect/querymgr/RealtimeQueryManager.kt
@@ -16,20 +16,25 @@
 
 package com.google.firebase.dataconnect.querymgr
 
-import com.google.firebase.dataconnect.DataSource
 import com.google.firebase.dataconnect.FirebaseDataConnect.CallerSdkType
 import com.google.firebase.dataconnect.QueryRef
 import com.google.firebase.dataconnect.core.DataConnectBidiConnectStream
+import com.google.firebase.dataconnect.core.DataConnectCache
 import com.google.firebase.dataconnect.core.DataConnectGrpcClient
 import com.google.firebase.dataconnect.core.DataConnectSerialization
 import com.google.firebase.dataconnect.core.Logger
 import com.google.firebase.dataconnect.core.LoggerGlobals.debug
+import com.google.firebase.dataconnect.core.QueryId
+import com.google.firebase.dataconnect.core.calculateQueryId
+import com.google.firebase.dataconnect.core.getEntityIdForPathFunction
+import com.google.firebase.dataconnect.sqlite.GetEntityIdForPathFunction
+import com.google.firebase.dataconnect.sqlite.SqliteSequencedReference
+import com.google.firebase.dataconnect.sqlite.map
 import com.google.firebase.dataconnect.util.CoroutineUtils.createChildSupervisorScope
 import com.google.firebase.dataconnect.util.IdStringGenerator
-import com.google.firebase.dataconnect.util.ImmutableByteArray
-import com.google.firebase.dataconnect.util.ProtoUtil.calculateSha512
 import com.google.firebase.dataconnect.util.update
 import com.google.protobuf.Struct
+import java.lang.System.currentTimeMillis
 import java.util.concurrent.atomic.AtomicReference
 import kotlinx.coroutines.CoroutineName
 import kotlinx.coroutines.CoroutineScope
@@ -53,6 +58,7 @@ internal class RealtimeQueryManager(
   coroutineScope: CoroutineScope,
   private val idStringGenerator: IdStringGenerator,
   private val serialization: DataConnectSerialization,
+  private val cache: DataConnectCache?,
   private val logger: Logger,
 ) {
 
@@ -85,7 +91,7 @@ internal class RealtimeQueryManager(
     callerSdkType: CallerSdkType,
     dataSerializersModule: SerializersModule?,
     variablesSerializersModule: SerializersModule?,
-  ): Flow<Result<Data>> {
+  ): Flow<Result<SqliteSequencedReference<Data>>> {
     val variablesStruct: Struct =
       serialization.encodeVariables(
         variables,
@@ -100,14 +106,16 @@ internal class RealtimeQueryManager(
         callerSdkType,
       )
 
-    return operationResultFlow.map {
+    return operationResultFlow.map { sourcedOperationResult ->
       runCatching {
-        serialization.decodeData(
-          it.data,
-          it.errors,
-          dataDeserializer,
-          dataSerializersModule,
-        )
+        sourcedOperationResult.map {
+          serialization.decodeData(
+            it.data,
+            it.errors,
+            dataDeserializer,
+            dataSerializersModule,
+          )
+        }
       }
     }
   }
@@ -116,7 +124,7 @@ internal class RealtimeQueryManager(
     operationName: String,
     variables: Struct,
     callerSdkType: CallerSdkType,
-  ): Flow<DataConnectGrpcClient.OperationResult> {
+  ): Flow<SqliteSequencedReference<DataConnectGrpcClient.OperationResult>> {
     val requestId = idStringGenerator.next("sub")
     val connection = ensureConnected(requestId, callerSdkType) ?: return emptyFlow()
 
@@ -134,25 +142,20 @@ internal class RealtimeQueryManager(
     requestId: String,
     operationName: String,
     variables: Struct,
-  ): Flow<DataConnectGrpcClient.OperationResult> {
-    // calculateSha512() is a CPU intensive operation that should NOT be performed on the main
+  ): Flow<SqliteSequencedReference<DataConnectGrpcClient.OperationResult>> {
+    // calculateQueryId() is a CPU intensive operation that should NOT be performed on the main
     // thread. This is the first reason why this method assumes it's running in this.coroutineScope.
-    val queryId = variables.calculateSha512(preamble = operationName)
+    val queryId = calculateQueryId(operationName, variables)
 
     // Acquiring the lock by an arbitrary thread could result in priority inversion. This is the
     // second reason why this method assumes it's running in this.coroutineScope: control over the
     // thread that acquires the lock.
     mutex.withLock {
       return flowByQueryId.getOrPut(queryId) {
-        val executeResponseFlow = stream.subscribe(requestId, operationName, variables)
-
-        executeResponseFlow.map { executeResponse ->
-          DataConnectGrpcClient.OperationResult(
-            data = executeResponse.data,
-            errors = executeResponse.errors,
-            source = DataSource.SERVER,
-          )
-        }
+        stream
+          .subscribe(requestId, operationName, variables)
+          .updateCache(cache, queryId)
+          .mapToOperationResponse()
       }
     }
   }
@@ -213,7 +216,7 @@ internal class RealtimeQueryManager(
     class Connected(val stream: DataConnectBidiConnectStream) : State {
       val mutex = Mutex()
       val flowByQueryId:
-        MutableMap<ImmutableByteArray, Flow<DataConnectGrpcClient.OperationResult>> =
+        MutableMap<QueryId, Flow<SqliteSequencedReference<DataConnectGrpcClient.OperationResult>>> =
         mutableMapOf()
       override fun toString() = "Connected"
     }
@@ -232,7 +235,7 @@ internal class RealtimeQueryManager(
 
 internal suspend fun <Data, Variables> RealtimeQueryManager.subscribe(
   queryRef: QueryRef<Data, Variables>
-): Flow<Result<Data>> =
+): Flow<Result<SqliteSequencedReference<Data>>> =
   subscribe(
     queryRef.operationName,
     queryRef.variables,
@@ -242,3 +245,47 @@ internal suspend fun <Data, Variables> RealtimeQueryManager.subscribe(
     queryRef.dataSerializersModule,
     queryRef.variablesSerializersModule,
   )
+
+private fun Flow<SqliteSequencedReference<DataConnectBidiConnectStream.ExecuteResponse>>
+  .mapToOperationResponse(): Flow<SqliteSequencedReference<DataConnectGrpcClient.OperationResult>> =
+  map {
+    SqliteSequencedReference(
+      it.sqliteSequenceNumber,
+      DataConnectGrpcClient.OperationResult(
+        data = it.ref.data,
+        errors = it.ref.errors,
+      ),
+    )
+  }
+
+private fun Flow<DataConnectBidiConnectStream.ExecuteResponse>.updateCache(
+  cache: DataConnectCache?,
+  queryId: QueryId,
+): Flow<SqliteSequencedReference<DataConnectBidiConnectStream.ExecuteResponse>> = map { response ->
+  val data = response.data
+  val sqliteSequenceNumber =
+    if (cache == null || data == null) {
+      null
+    } else {
+      cache
+        .open()
+        .insertQueryResult(
+          response.authUid,
+          queryId,
+          data,
+          cache.maxAgeProto,
+          currentTimeMillis(),
+          response.getEntityIdForPathFunction(),
+        )
+    }
+  SqliteSequencedReference(sqliteSequenceNumber, response)
+}
+
+@JvmName("getEntityIdForPathFunction_DataConnectBidiConnectStream_ExecuteResponse")
+private fun DataConnectBidiConnectStream.ExecuteResponse.getEntityIdForPathFunction():
+  GetEntityIdForPathFunction? =
+  if (extensions.isEmpty()) {
+    null
+  } else {
+    extensions.getEntityIdForPathFunction()
+  }