+++ firebase-appdistribution-gradle/CHANGELOG.md
@@ -1,3 +1,7 @@
+### Unreleased
+
+* {{feature}} Added authentication support for Application Default Credentials and Workload Identity Federation (#6936)
+
 ### App Distribution Gradle plugin version 5.2.1
 
 * {{fixed}} Declare dependency on GoogleServices plugin